+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/morefeatures/MoreFeaturesScreen.kt
@@ -44,7 +44,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import br.com.rbrthmn.R
 import br.com.rbrthmn.ui.financialcompanion.navigation.NavigationDestination
 import br.com.rbrthmn.ui.financialcompanion.screens.recurringexpenses.RecurringExpensesDestination
-import br.com.rbrthmn.ui.financialcompanion.screens.ReservesDestination
+import br.com.rbrthmn.ui.financialcompanion.screens.reserves.ReservesDestination
 import br.com.rbrthmn.ui.financialcompanion.screens.SettingsDestination
 import br.com.rbrthmn.ui.financialcompanion.screens.incomedivisions.IncomeDivisionsDestination
